requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
asn1crypto==1.5.1
    # via snowflake-connector-python
boto3==1.38.27
    # via
    #   -r requirements.in
    #   snowflake-connector-python
botocore==1.38.27
    # via
    #   boto3
    #   s3transfer
    #   snowflake-connector-python
certifi==2025.4.26
    # via
    #   requests
    #   snowflake-connector-python
cffi==1.17.1
    # via
    #   cryptography
    #   snowflake-connector-python
charset-normalizer==3.4.2
    # via
    #   requests
    #   snowflake-connector-python
cryptography==45.0.3
    # via
    #   pyopenssl
    #   snowflake-connector-python
filelock==3.18.0
    # via snowflake-connector-python
idna==3.10
    # via
    #   requests
    #   snowflake-connector-python
iniconfig==2.1.0
    # via pytest
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
numpy==2.2.6
    # via pandas
packaging==25.0
    # via
    #   pytest
    #   snowflake-connector-python
pandas==2.2.3
    # via -r requirements.in
platformdirs==4.3.8
    # via snowflake-connector-python
pluggy==1.6.0
    # via pytest
pyarrow<19.0.0
    # via -r requirements.in
pycparser==2.22
    # via cffi
pyjwt==2.10.1
    # via snowflake-connector-python
pyopenssl==25.1.0
    # via snowflake-connector-python
pytest==8.3.5
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via
    #   botocore
    #   pandas
python-dotenv==1.1.0
    # via -r requirements.in
pytz==2025.2
    # via
    #   pandas
    #   snowflake-connector-python
requests==2.32.3
    # via
    #   -r requirements.in
    #   snowflake-connector-python
s3transfer==0.13.0
    # via boto3
six==1.17.0
    # via python-dateutil
snowflake-connector-python==3.15.0
    # via -r requirements.in
sortedcontainers==2.4.0
    # via snowflake-connector-python
tomlkit==0.13.2
    # via snowflake-connector-python
typing-extensions==4.14.0
    # via
    #   pyopenssl
    #   snowflake-connector-python
tzdata==2025.2
    # via pandas
urllib3==2.4.0
    # via
    #   botocore
    #   requests
